+++ data-raw/cabat-dict.R
@@ -12,5 +12,10 @@ names(spanish)[[1]] <- "key"
 stopifnot(all(input$key == spanish$key))
 input$ES <- spanish$ES
 
+dutch <- read.csv("data-raw/dict-dutch.csv", stringsAsFactors = FALSE)
+names(dutch)[[1]] <- "key"
+stopifnot(all(input$key == dutch$key))
+input$NL <- dutch$NL
+
 cabat_dict <- psychTestR::i18n_dict$new(input)
 usethis::use_data(cabat_dict, overwrite = TRUE)